StaticLists

Documentation for StaticLists.

StaticLists.KeyedStaticList

StaticLists.StaticList

StaticLists.deleteat

StaticLists.pop

StaticLists.popat

StaticLists.popfirst

StaticLists.push

StaticLists.pushfirst

StaticLists.KeyedStaticList — Type

KeyedStaticList(items::Pair...)
KeyedStaticList(keys::StaticList, values::StaticList)

An instance of StaticList with keys for each element.

StaticLists.StaticList — Type

StaticList(items...)

A statically-sized, singly-linked StaticList.

StaticLists.deleteat — Method

StaticLists.deleteat(StaticList, key)

Returns a StaticList without the value corresponding to key.

Warning

This is not part of the public API and may change without notice.

StaticLists.pop — Method

StaticLists.pop(StaticList) -> (last(StaticList), Base.front(StaticList))

Returns a tuple with the last item and the StaticList without the last item.

Warning

This is not part of the public API and may change without notice.

StaticLists.popat — Method

StaticLists.popat(StaticList, key) -> (StaticList[key], StaticLists.delete(StaticList, key))

Returns the value at key and the StaticList without the value.

Warning

This is not part of the public API and may change without notice.

StaticLists.popfirst — Method

StaticLists.popfirst(StaticList) -> (first(StaticList), Base.tail(StaticList))

Returns a tuple with the first item and the StaticList without the first item.

Warning

This is not part of the public API and may change without notice.

StaticLists.push — Method

push(StaticList, item)

Returns a new StaticList with item added to the end.

StaticLists.pushfirst — Method

pushfirst(StaticList, item)

Returns a new StaticList with item added to the front.