StaticLists

Documentation for StaticLists.

StaticLists.KeyedList

StaticLists.List

StaticLists.pop

StaticLists.popat

StaticLists.popfirst

StaticLists.push

StaticLists.pushfirst

StaticLists.KeyedList — Type

KeyedList(items::Pair...)
KeyedList(keys::List, values::List)

An instance of List with keys for each element.

StaticLists.List — Type

List(items...)

A statically sized, singly linked list.

StaticLists.pop — Method

StaticLists.pop(list) -> (last(list), Base.front(list))

Returns a tuple with the last item and the list without the last item.

StaticLists.popat — Method

StaticLists.popat(list, key) -> (list[key], StaticLists.delete(list, key))

Returns the value at key and the list without the value.

StaticLists.popfirst — Method

StaticLists.popfirst(list) -> (first(list), Base.tail(list))

Returns a tuple with the first item and the list without the first item.

StaticLists.push — Method

StaticLists.push(list, item)

Returns a new list with item added to the end.

StaticLists.pushfirst — Method

StaticLists.pushfirst(list, item)

Returns a new list with item added to the front.